Tmp/socket keyboard refactoring (#19)
* refactoring
